# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Interview Reader API
  version: 0.2.0
  description: Reader study APIs and isolated desktop administration APIs.
servers:
  - url: /api
security:
  - cookieAuth: []
paths:
...
  /reader/versions/{versionId}/toc:
    get:
      summary: Get the published table of contents
      parameters:
        - $ref: "#/components/parameters/VersionId"
      responses:
        default: { $ref: "#/components/responses/Problem" }
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/TocNode" }
  /reader/versions/{versionId}/nodes/{nodeId}/content:
    get:
      summary: Get a reading node and its content blocks
      parameters:
        - $ref: "#/components/parameters/VersionId"
        - $ref: "#/components/parameters/NodeId"
        - { name: afterSeq, in: query, schema: { type: integer } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 50 } }
      responses:
        default: { $ref: "#/components/responses/Problem" }
        "200":
          description: Successful response
          content:
            application/json:
              schema: { $ref: "#/components/schemas/NodeContent" }
...
  /admin/import-jobs:
    post:
      summary: Create an import job for a new document or target an existing document
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: PDF, XLSX, Markdown, or JSON package, up to 10 MiB; the server detects the source type.
                targetDocumentId: { type: string, format: uuid }
      responses:
        default: { $ref: "#/components/responses/Problem" }
        "202":
          description: Successful response
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ImportJob" }
  /admin/import-jobs/{jobId}:
    get:
      summary: Get import progress and terminal status
      parameters:
        - $ref: "#/components/parameters/JobId"
      responses:
        default: { $ref: "#/components/responses/Problem" }
        "200":
          description: Successful response
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ImportJob" }
  /admin/import-jobs/{jobId}/issues:
    get:
      summary: List import validation issues
      parameters:
        - $ref: "#/components/parameters/JobId"
      responses:
        default: { $ref: "#/components/responses/Problem" }
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ImportIssue" }
  /admin/import-jobs/{jobId}/document-metadata:
    get:
      summary: Preview staged document metadata and any exact code match
      parameters:
        - $ref: "#/components/parameters/JobId"
      responses:
        default: { $ref: "#/components/responses/Problem" }
        "200":
          description: Successful response
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ImportDocumentPreview" }
    patch:
      summary: Revise staged metadata for a new document; existing targets remain read-only
      parameters:
        - $ref: "#/components/parameters/JobId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateImportDocumentMetadataRequest" }
      responses:
        default: { $ref: "#/components/responses/Problem" }
        "200":
          description: Staged metadata updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ImportDocumentPreview" }
        "409": { $ref: "#/components/responses/Problem" }
  /admin/import-jobs/{jobId}/raw-extraction:
    get:
      summary: Get converter extraction data for diagnostics
      parameters:
        - $ref: "#/components/parameters/JobId"
      responses:
        default: { $ref: "#/components/responses/Problem" }
        "200":
          description: Successful response
          content:
            application/json:
              schema: { $ref: "#/components/schemas/JsonValue" }
  /admin/import-jobs/{jobId}/source-file:
    get:
      summary: Get original uploaded source file
      parameters:
        - $ref: "#/components/parameters/JobId"
      responses:
        default: { $ref: "#/components/responses/Problem" }
        "200":
          description: Original source file
          content:
            application/pdf:
              schema: { type: string, format: binary }
            application/octet-stream:
              schema: { type: string, format: binary }
  /admin/import-jobs/{jobId}/normalized-package:
    get:
      summary: Get the staged document package
      parameters:
        - $ref: "#/components/parameters/JobId"
      responses:
        default: { $ref: "#/components/responses/Problem" }
        "200":
          description: Successful response
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DocumentPackage" }
  /admin/import-jobs/{jobId}/normalized-package/sections/{sectionKey}:
    patch:
      summary: Revise a staged section before commit
      parameters:
        - $ref: "#/components/parameters/JobId"
        - { name: sectionKey, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/JsonObject" }
      responses:
        default: { $ref: "#/components/responses/Problem" }
        "200":
          description: Successful response
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DocumentPackage" }
  /admin/import-jobs/{jobId}/normalized-package/blocks/{blockKey}:
    patch:
      summary: Revise a staged content block before commit
      parameters:
        - $ref: "#/components/parameters/JobId"
        - { name: blockKey, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/JsonObject" }
      responses:
        default: { $ref: "#/components/responses/Problem" }
        "200":
          description: Successful response
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DocumentPackage" }
  /admin/import-jobs/{jobId}/commit:
    post:
      summary: Commit a validated package as a draft version
      parameters:
        - $ref: "#/components/parameters/JobId"
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CommitImportRequest" }
      responses:
        default: { $ref: "#/components/responses/Problem" }
        "201":
          description: Successful response
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DocumentVersion" }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: IR_SESSION
  parameters:
...
    VersionId: { name: versionId, in: path, required: true, schema: { type: string, format: uuid } }
    NodeId: { name: nodeId, in: path, required: true, schema: { type: string, format: uuid } }
    JobId: { name: jobId, in: path, required: true, schema: { type: string, format: uuid } }
  responses:
...
    Problem:
      description: RFC 9457-compatible problem detail
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
  schemas:
    JsonValue: {}
    JsonObject:
      type: object
      additionalProperties: true
    Uuid:
      type: string
      format: uuid
    NullableUuid:
      type: [string, "null"]
      format: uuid
    SourceType:
      type: string
      enum: [PDF, EXCEL, JSON_PACKAGE, MARKDOWN, MANUAL]
    NodeType:
      type: string
      enum: [PART, CHAPTER, SECTION, SUBSECTION, QUESTION, APPENDIX, OTHER]
    BlockType:
      type: string
      enum: [paragraph, heading_note, unordered_list, ordered_list, code, table, quote, callout, formula, image, divider, table_snapshot]
    SemanticRole:
      type: string
      enum: [QUESTION, ANSWER, EXPLANATION, CONCLUSION, INTRODUCTION, DIRECTORY, PRINCIPLE, PRACTICE, PITFALL, FOLLOW_UP]
    VersionStatus:
      type: string
      enum: [DRAFT, PUBLISHED, RETIRED]
    DocumentStatus:
      type: string
      enum: [DRAFT, PUBLISHED, OFFLINE, DELETING, DELETE_FAILED]
...
    ImportStatus:
      type: string
      enum: [UPLOADED, PREFLIGHT, EXTRACTING, NORMALIZING, VALIDATING, READY, REVIEW_REQUIRED, IMPORTED, FAILED, CANCELED]
    ImportStage:
      type: string
      enum: [UPLOADED, PREFLIGHT, EXTRACTING, NORMALIZING, VALIDATING, READY, REVIEW_REQUIRED, REVIEWING, FAILED, CANCELED, COMMITTED, DRAFT_DISCARDED]
...
    Problem:
      type: object
      required: [type, title, status, detail, code, error, traceId]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer, minimum: 400, maximum: 599 }
        detail: { type: string }
        instance: { type: string, format: uri }
        code: { type: string }
        error: { type: string }
        traceId: { type: string, format: uuid }
        fieldErrors:
          type: object
          additionalProperties: { type: string }
...
    TocNode:
      type: object
      required: [id, parentId, title, level, nodeType, semanticRole, anchor, sourcePageStart, children]
      properties:
        id: { $ref: "#/components/schemas/Uuid" }
        parentId: { $ref: "#/components/schemas/NullableUuid" }
        title: { type: string }
        level: { type: integer, minimum: 1 }
        nodeType: { $ref: "#/components/schemas/NodeType" }
        semanticRole:
          anyOf: [{ $ref: "#/components/schemas/SemanticRole" }, { type: "null" }]
        anchor: { type: string }
        sourcePageStart: { type: [integer, "null"], minimum: 1 }
        children: { type: array, items: { $ref: "#/components/schemas/TocNode" } }
    SourceBbox:
      type: object
      additionalProperties: true
      properties:
        page: { type: [integer, "null"] }
        x: { type: [number, "null"] }
        y: { type: [number, "null"] }
        width: { type: [number, "null"] }
        height: { type: [number, "null"] }
    ContentBlock:
      type: object
      required: [id, blockKey, seq, blockType, payload, plainText, sourcePage, sourceBbox, confidence]
      properties:
        id: { $ref: "#/components/schemas/Uuid" }
        blockKey: { type: string }
        seq: { type: integer }
        blockType: { $ref: "#/components/schemas/BlockType" }
        payload: { $ref: "#/components/schemas/JsonObject" }
        plainText: { type: string }
        sourcePage: { type: [integer, "null"] }
        sourceBbox:
          anyOf: [{ $ref: "#/components/schemas/SourceBbox" }, { type: "null" }]
        confidence: { type: [number, "null"], minimum: 0, maximum: 1 }
    NodeContent:
      type: object
      required: [node, blocks, nextAfterSeq]
      properties:
        node: { $ref: "#/components/schemas/TocNode" }
        blocks: { type: array, items: { $ref: "#/components/schemas/ContentBlock" } }
        nextAfterSeq: { type: [integer, "null"] }
...
    DocumentInfo:
      type: object
      required: [documentKey, title, description, language, tags]
      properties:
        documentKey: { type: string }
        title: { type: string }
        description: { type: [string, "null"] }
        language: { type: string }
        tags: { type: array, items: { type: string } }
    VersionInfo:
      type: object
      required: [versionKey, sourceType, sourceFileName, sourceSha256, converterVersion, metadata]
      properties:
        versionKey: { type: string }
        sourceType: { $ref: "#/components/schemas/SourceType" }
        sourceFileName: { type: [string, "null"] }
        sourceSha256: { type: [string, "null"], pattern: "^[0-9a-fA-F]{64}$" }
        converterVersion: { type: [string, "null"] }
        metadata: { $ref: "#/components/schemas/JsonObject" }
    DocumentPackageSection:
      type: object
      required: [sectionKey, parentSectionKey, level, nodeType, semanticRole, title, sortOrder, anchor, sourcePageStart, sourcePageEnd, sourceBbox, contentHash]
      properties:
        sectionKey: { type: string }
        parentSectionKey: { type: [string, "null"] }
        level: { type: integer, minimum: 1 }
        nodeType: { $ref: "#/components/schemas/NodeType" }
        semanticRole:
          anyOf: [{ $ref: "#/components/schemas/SemanticRole" }, { type: "null" }]
        title: { type: string }
        sortOrder: { type: integer }
        anchor:
          type: string
          description: Server-generated stable internal identifier. Supplied values are ignored during import.
        sourcePageStart: { type: [integer, "null"] }
        sourcePageEnd: { type: [integer, "null"] }
        sourceBbox:
          anyOf: [{ $ref: "#/components/schemas/SourceBbox" }, { type: "null" }]
        contentHash: { type: [string, "null"] }
    DocumentPackageBlock:
      type: object
      required: [blockKey, sectionKey, seq, blockType, payload, plainText, language, sourcePage, sourceBbox, confidence, contentHash]
      properties:
        blockKey: { type: string }
        sectionKey: { type: string }
        seq: { type: integer }
        blockType: { $ref: "#/components/schemas/BlockType" }
        payload: { $ref: "#/components/schemas/JsonObject" }
        plainText: { type: string }
        language: { type: [string, "null"] }
        sourcePage: { type: [integer, "null"] }
        sourceBbox:
          anyOf: [{ $ref: "#/components/schemas/SourceBbox" }, { type: "null" }]
        confidence: { type: [number, "null"], minimum: 0, maximum: 1 }
        contentHash: { type: [string, "null"] }
    AssetInfo:
      type: object
      required: [assetKey, path, mimeType, sha256, alt]
      properties:
        assetKey: { type: string }
        path: { type: string }
        mimeType: { type: string }
        sha256: { type: string }
        alt: { type: [string, "null"] }
    DocumentPackage:
      type: object
      required: [schemaVersion, document, version, sections, blocks, assets]
      properties:
        schemaVersion: { type: string, enum: ["1.0"] }
        document: { $ref: "#/components/schemas/DocumentInfo" }
        version: { $ref: "#/components/schemas/VersionInfo" }
        sections: { type: array, items: { $ref: "#/components/schemas/DocumentPackageSection" } }
        blocks: { type: array, items: { $ref: "#/components/schemas/DocumentPackageBlock" } }
        assets: { type: array, items: { $ref: "#/components/schemas/AssetInfo" } }
    ImportJob:
      type: object
      required: [id, targetDocumentId, sourceType, status, currentStage, progress, statistics, errorCode, errorMessage]
      properties:
        id: { $ref: "#/components/schemas/Uuid" }
        targetDocumentId: { $ref: "#/components/schemas/NullableUuid" }
        sourceType: { $ref: "#/components/schemas/SourceType" }
        status: { $ref: "#/components/schemas/ImportStatus" }
        currentStage:
          anyOf: [{ $ref: "#/components/schemas/ImportStage" }, { type: "null" }]
        progress: { type: integer, minimum: 0, maximum: 100 }
        statistics: { $ref: "#/components/schemas/JsonObject" }
        errorCode: { type: [string, "null"] }
        errorMessage: { type: [string, "null"] }
    ImportResolution:
      type: string
      enum: [CREATE_NEW, IMPORT_AS_NEW_VERSION]
    ImportIssueSeverity:
      type: string
      enum: [BLOCKING, WARNING]
    ExistingDocumentMatch:
      type: object
      required: [id, code, title, status]
      properties:
        id: { $ref: "#/components/schemas/Uuid" }
        code: { type: string }
        title: { type: string }
        status: { $ref: "#/components/schemas/DocumentStatus" }
    ImportDocumentPreview:
      type: object
      required: [documentKey, title, description, tags, editable, matchingDocument, suggestedDocumentKey, duplicateTitleCount]
      properties:
        documentKey: { type: string, readOnly: true }
        title: { type: string, minLength: 1, maxLength: 500 }
        description: { type: [string, "null"], maxLength: 5000 }
        tags:
          type: array
          maxItems: 20
          items: { type: string, minLength: 1, maxLength: 50 }
        editable: { type: boolean }
        matchingDocument:
          anyOf: [{ $ref: "#/components/schemas/ExistingDocumentMatch" }, { type: "null" }]
        suggestedDocumentKey: { type: [string, "null"], readOnly: true }
        duplicateTitleCount: { type: integer, minimum: 0 }
    UpdateImportDocumentMetadataRequest:
      type: object
      required: [title, description, tags]
      properties:
        title: { type: string, minLength: 1, maxLength: 500 }
        description: { type: [string, "null"], maxLength: 5000 }
        tags:
          type: array
          maxItems: 20
          items: { type: string, minLength: 1, maxLength: 50 }
    CommitImportRequest:
      type: object
      required: [resolution]
      properties:
        resolution: { $ref: "#/components/schemas/ImportResolution" }
    ImportIssue:
      type: object
      required: [severity, issueCode, message, sourcePage, sectionKey, blockKey, cellRef]
      properties:
        severity: { $ref: "#/components/schemas/ImportIssueSeverity" }
        issueCode: { type: string }
        message: { type: string }
        sourcePage: { type: [integer, "null"] }
        sectionKey: { type: [string, "null"] }
        blockKey: { type: [string, "null"] }
        cellRef: { type: [string, "null"] }
    DocumentVersion:
      type: object
      required: [id, documentId, versionNo, status, schemaVersion]
      properties:
        id: { $ref: "#/components/schemas/Uuid" }
        documentId: { $ref: "#/components/schemas/Uuid" }
        versionNo: { type: integer, minimum: 1 }
        status: { $ref: "#/components/schemas/VersionStatus" }
        schemaVersion: { type: string }
...